% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{access_log}
\alias{access_log}
\title{Access the log of actions of this package}
\usage{
access_log()
}
\value{
A dribble, containing the info for the log googlesheet
}
\description{
Access the log of actions of this package
}
