% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_docs.R
\name{add_status}
\alias{add_status}
\title{Create a status row}
\usage{
add_status(df, string, pattern, task, signoff)
}
\arguments{
\item{df}{data.frame of status to be appended to}

\item{string}{passed to \link[stringr]{str_extract}}

\item{pattern}{passed to \link[stringr]{str_extract}}

\item{task}{name of the task in the new status row}

\item{signoff}{name of the signoff in the new status row}
}
\value{
a data.frame
}
\description{
Create a status row
}
