% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_statuses.R
\name{as_statuses}
\alias{as_statuses}
\title{Coerce into type \code{statuses}}
\usage{
as_statuses(x)
}
\arguments{
\item{x}{object to be coerced, probably a data.frame}
}
\value{
object of type \code{statuses}
}
\description{
Coerce into type \code{statuses}
}
