% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_drive.R
\name{check_files_roadmap}
\alias{check_files_roadmap}
\title{Verify roadmap files against central list of}
\usage{
check_files_roadmap(
  db_units = read_db_units(),
  roadmap_files = find_files_roadmap()
)
}
\arguments{
\item{db_units}{the tibble with the central DB of all units}

\item{roadmap_files}{the dribble with all files matching "Unit Roadmap" in the filename}
}
\description{
Verify roadmap files against central list of
}
