% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_docs.R
\name{check_roadmap_files}
\alias{check_roadmap_files}
\title{Verify roadmap files against central list of}
\usage{
check_roadmap_files(
  db_units = read_db_units(),
  roadmap_files = find_roadmap_files()
)
}
\arguments{
\item{db_units}{the tibble with the central DB of all units}

\item{roadmap_files}{the dribble with all files matching "Unit Roadmap" in the filename}
}
\description{
Verify roadmap files against central list of
}
