% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_docs.R
\name{check_signoff_doc_index}
\alias{check_signoff_doc_index}
\title{Check That Doc Index for Sign-Off is in Range}
\usage{
check_signoff_doc_index(doc_index, phase_labels, signoff_labels, phase)
}
\arguments{
\item{doc_index}{index for the paragraph with the signoff}

\item{phase_labels}{a data.frame, subset of the output from \link[officer]{docx_summary} that are the paragraphs corresponding to the labels of the start of each phase}

\item{signoff_labels}{a data.frame, subset of the output from \link[officer]{docx_summary} that are the paragraphs corresponding to the "Sign offs" headings}

\item{phase}{which phase to check}
}
\description{
Check That Doc Index for Sign-Off is in Range
}
