% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_statuses.R
\name{check_statuses}
\alias{check_statuses}
\title{Verify that x is an object of type \code{statuses} or is coerceable}
\usage{
check_statuses(x)
}
\arguments{
\item{x}{an object of type \code{statuses} or a data.frame}
}
\value{
logical
}
\description{
Verify that x is an object of type \code{statuses} or is coerceable
}
