% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-functions.R
\name{decrypt_gdrive_token}
\alias{decrypt_gdrive_token}
\title{Get a Token for access to google drive}
\usage{
decrypt_gdrive_token(
  token_RDS = ".secrets/gdrive-token.rds",
  decrypt_env_var = "GDRIVE_KEY"
)
}
\value{
an OAuth token
}
\description{
This function expects that there is an encrypted file in RDS format with an
OAuth token, and an environmental variable with the key to decrypt the token.
}
