% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_drive.R
\name{download_unit_data}
\alias{download_unit_data}
\title{Download Data on Units and Statuses}
\usage{
download_unit_data(data_file = "unit_data.RDS")
}
\arguments{
\item{data_file}{path to the desired download location}
}
\value{
a list with components for the statuses, units, and timestamp
}
\description{
Download Data on Units and Statuses
}
