% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_docs.R
\name{extract_roadmap_miniunits}
\alias{extract_roadmap_miniunits}
\title{Extract the mini-unit titles of a roadmap}
\usage{
extract_roadmap_miniunits(content)
}
\arguments{
\item{content}{a data.frame (output from \link[officer]{docx_summary})}
}
\value{
character vector (of length 8)
}
\description{
Extract the mini-unit titles of a roadmap
}
