% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_docs.R
\name{extract_roadmap_phases}
\alias{extract_roadmap_phases}
\title{Extract the labels for each Phase}
\usage{
extract_roadmap_phases(content, NUM_PHASES = 7)
}
\arguments{
\item{content}{a data.frame (output from \link[officer]{docx_summary})}

\item{NUM_PHASES}{number of phases to expect}
}
\value{
a data.frame
}
\description{
Extract the labels for each Phase
}
