% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_docs.R
\name{extract_roadmap_signoffs}
\alias{extract_roadmap_signoffs}
\title{Extract the signoff sections}
\usage{
extract_roadmap_signoffs(content)
}
\arguments{
\item{content}{a data.frame (output from \link[officer]{docx_summary})}
}
\value{
a data.frame
}
\description{
Extract the signoff sections
}
