% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_docs.R
\name{extract_roadmap_statuses}
\alias{extract_roadmap_statuses}
\title{Extract roadmap statuses}
\usage{
extract_roadmap_statuses(content)
}
\arguments{
\item{content}{a data.frame (output from \link[officer]{docx_summary})}
}
\value{
a data.frame
}
\description{
Extract roadmap statuses
}
