% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_docs.R
\name{extract_roadmap_title}
\alias{extract_roadmap_title}
\title{Extract the title of a roadmap}
\usage{
extract_roadmap_title(content)
}
\arguments{
\item{content}{a data.frame (output from \link[officer]{docx_summary})}
}
\value{
character
}
\description{
Extract the title of a roadmap
}
