% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_docs.R
\name{find_roadmap_phases}
\alias{find_roadmap_phases}
\title{Find the labels for each Phase}
\usage{
find_roadmap_phases(content, NUM_PHASES = 7)
}
\arguments{
\item{content}{a data.frame (output from \link[officer]{docx_summary})}

\item{NUM_PHASES}{number of phases to expect}
}
\value{
a data.frame
}
\description{
Find the labels for each Phase
}
