% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_docs.R
\name{find_roadmap_signoffs}
\alias{find_roadmap_signoffs}
\title{Find the signoff sections}
\usage{
find_roadmap_signoffs(content)
}
\arguments{
\item{content}{a data.frame (output from \link[officer]{docx_summary})}
}
\value{
a data.frame
}
\description{
Find the signoff sections
}
