% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_docs.R
\name{find_status}
\alias{find_status}
\title{Create a status row}
\usage{
find_status(string, pattern, phase, task, signoff)
}
\arguments{
\item{string}{passed to \link[stringr]{str_extract}}

\item{pattern}{passed to \link[stringr]{str_extract}}

\item{phase}{number of the phase in the new status row}

\item{task}{name of the task in the new status row}

\item{signoff}{name of the signoff in the new status row}
}
\value{
a data.frame
}
\description{
Create a status row
}
