% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_docs.R
\name{find_status_phase_1}
\alias{find_status_phase_1}
\alias{find_status_phase_2}
\title{Find Phase 1 Statuses}
\usage{
find_status_phase_1(table_cells, phase_labels, signoff_labels)

find_status_phase_2(table_cells, phase_labels, signoff_labels)
}
\arguments{
\item{table_cells}{a data.frame, subset of the output from \link[officer]{docx_summary} that are the table cells}

\item{phase_labels}{a data.frame, subset of the output from \link[officer]{docx_summary} that are the paragraphs corresponding to the labels of the start of each phase}

\item{signoff_labels}{a data.frame, subset of the output from \link[officer]{docx_summary} that are the paragraphs corresponding to the "Sign offs" headings}
}
\value{
a data.frame
}
\description{
Find Phase 1 Statuses
}
\section{Functions}{
\itemize{
\item \code{find_status_phase_2()}: Find Phase 2 Statuses

}}
