% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_docs.R
\name{find_statuses}
\alias{find_statuses}
\title{Find Statuses}
\usage{
find_statuses(content)
}
\arguments{
\item{content}{a data.frame (output from \link[officer]{docx_summary})}
}
\value{
a data.frame
}
\description{
Find Statuses
}
