% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{fix_loop_num}
\alias{fix_loop_num}
\title{Fix the loop_num using the stored value or set to NA}
\usage{
fix_loop_num(loop_num)
}
\arguments{
\item{loop_num}{number indicating which processing loop is active}
}
\value{
a number, the current loop number
}
\description{
Fix the loop_num using the stored value or set to NA
}
