% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications.R
\name{format_status_msg}
\alias{format_status_msg}
\title{Format a status row for printing}
\usage{
format_status_msg(status_row, sep = "\\n")
}
\arguments{
\item{status_row}{a single row of a \code{statuses} object}

\item{sep}{character to separate each element in formatted output}
}
\value{
character (formatted message)
}
\description{
Format a status row for printing
}
