% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_docs.R
\name{format_statuses}
\alias{format_statuses}
\title{Cleanup and Format Statuses}
\usage{
format_statuses(statuses, title, mini_unit_names)
}
\arguments{
\item{statuses}{a data.frame, usually the output of \link{extract_roadmap_statuses}}

\item{title}{name of the unit}

\item{mini_unit_names}{a character vector, usually the output of \link{extract_roadmap_mini_units}}
}
\value{
a data.frame
}
\description{
Cleanup and Format Statuses
}
