% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrive.automation-package.R
\docType{package}
\name{gdrive.automation-package}
\alias{gdrive.automation}
\alias{gdrive.automation-package}
\title{gdrive.automation: Sync and Automate C4R Unit Progress}
\description{
The goal of gdrive-automation is to automate processing and tasks related to the C4R Shared Drive, to aid in synchronization of states and content.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://c4r-io.github.io/gdrive-automation/}
  \item Report bugs at \url{https://github.com/c4r-io/gdrive-automation/issues}
}

}
\author{
\strong{Maintainer}: Hao Ye \email{haoye@seas.upenn.edu} (\href{https://orcid.org/0000-0002-8630-1458}{ORCID})

}
\keyword{internal}
