% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_monday_token}
\alias{get_monday_token}
\title{Get a token for access to Monday}
\usage{
get_monday_token(token_env_var = "monday_token")
}
\arguments{
\item{token_env_var}{name of the environmental variable that is the token}
}
\value{
a character
}
\description{
Get a token for access to Monday
}
