% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_docs.R
\name{get_roadmap_statuses}
\alias{get_roadmap_statuses}
\title{Download a roadmap and extract statuses}
\usage{
get_roadmap_statuses(roadmap_id, dl_path = tempfile(fileext = ".docx"))
}
\arguments{
\item{roadmap_id}{the drv_id of a roadmap file}

\item{dl_path}{path to where to download the roadmap file}
}
\value{
a data.frame of statuses, mirroring the structure of \link{read_unit_tasks}
}
\description{
Download a roadmap and extract statuses
}
