% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications.R
\name{get_users}
\alias{get_users}
\title{Query Monday for user emails and ids}
\usage{
get_users(token = get_monday_token())
}
\value{
data.frame (columns: \code{email} and \code{id})
}
\description{
\inheritParams{post_query}
}
