% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_statuses.R
\name{handle_diff_statuses}
\alias{handle_diff_statuses}
\title{Handle Differences in Roadmap and Tracker Statuses}
\usage{
handle_diff_statuses(roadmap_dat, roadmap_url, tracker_dat, tracker_url)
}
\arguments{
\item{roadmap_dat}{object of type \code{statuses} (from Roadmap)}

\item{roadmap_url}{character (URL of roadmap)}

\item{tracker_dat}{object of type \code{statuses} (from tracker)}

\item{tracker_url}{character (URL of tracker)}
}
\value{
object of type \code{statuses}
}
\description{
Handle Differences in Roadmap and Tracker Statuses
}
