% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_statuses.R
\name{identify_todo}
\alias{identify_todo}
\title{Identify the next action item for a \code{statuses} object}
\usage{
identify_todo(x)
}
\arguments{
\item{x}{object of type \code{statuses}}
}
\value{
character (what next action item is)
}
\description{
Identify the next action item for a \code{statuses} object
}
