% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_statuses.R
\name{is.statuses}
\alias{is.statuses}
\title{Check if object is of type \code{statuses}}
\usage{
is.statuses(x)
}
\arguments{
\item{x}{object to check}
}
\value{
logical
}
\description{
Check if object is of type \code{statuses}
}
