% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_action}
\alias{log_action}
\title{Record an action in the log}
\usage{
log_action(action = "", note = "", type = "INFO", loop_num = NA)
}
\arguments{
\item{action}{Description of the change}

\item{note}{Any relevant notes (e.g. if the change is part of a larger set
of changes)}

\item{type}{Type of message ("INFO", "ERROR", "ACTION TAKEN", "ACTION NEEDED")}

\item{loop_num}{Number indicating which processing loop the log message is associated with}
}
\description{
All changes to files in the C4R googledrive that are performed by this
package should be accompanied with a log entry, added via this function.
}
