% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_status_update}
\alias{log_status_update}
\title{generate status update content}
\usage{
log_status_update(status_msg, tracker_url)
}
\arguments{
\item{status_msg}{formatted status msg}

\item{tracker_url}{URL of a google sheet (Unit Tracker)}
}
\description{
generate status update content
}
