% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications.R
\name{match_users}
\alias{match_users}
\title{Match user name or email to user_id}
\usage{
match_users(to = "Hao Ye")
}
\arguments{
\item{to}{vector of names or emails}
}
\value{
numeric ids
}
\description{
Match user name or email to user_id
}
