% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_statuses.R
\name{mini_units<-}
\alias{mini_units<-}
\title{Set mini-unit titles of \code{statuses} object}
\usage{
mini_units(x) <- value
}
\arguments{
\item{x}{object of type \code{statuses}}

\item{value}{new mini-unit titles (character)}
}
\value{
object of type \code{statuses}
}
\description{
Set mini-unit titles of \code{statuses} object
}
