% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_statuses.R
\name{mini_units}
\alias{mini_units}
\title{Get mini-unit titles of \code{statuses} object}
\usage{
mini_units(x)
}
\arguments{
\item{x}{object of type \code{statuses}}
}
\value{
character (length = of mini-units)
}
\description{
Get mini-unit titles of \code{statuses} object
}
