% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications.R
\name{msg}
\alias{msg}
\alias{format_status_msg}
\alias{format_notification_msg}
\title{Format a status row for printing}
\usage{
format_status_msg(status_row, sep = "\\n")

format_notification_msg(prefix, status_row, sep = "\\n")
}
\arguments{
\item{status_row}{a single row of a \code{statuses} object}

\item{sep}{character to separate each element in formatted output}

\item{prefix}{optional string to put in front of message}
}
\value{
character (formatted message)
}
\description{
Format a status row for printing
}
