% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications.R
\name{notify}
\alias{notify}
\title{Send notification}
\usage{
notify(to = "CENTER", status_row)
}
\arguments{
\item{to}{either "CENTER" or "NIH" or a specific METER}

\item{status_row}{a single row of a \code{statuses} object}
}
\description{
Send notification
}
