% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrive.automation-package.R
\docType{data}
\name{parsing_dat}
\alias{parsing_dat}
\title{Expected Structure of Statuses}
\format{
\subsection{\code{who}}{

A data frame with 22 rows and 5 columns:
\describe{
\item{phase}{Unit Roadmap Phase}
\item{parse_group}{Identify which statuses occur in a shared signoff box}
\item{task}{name of the task}
\item{signoff}{name of the party responsible for signoff}
\item{label}{string that identifies the status}
}
}
}
\usage{
parsing_dat
}
\description{
Expected Structure of Statuses
}
\keyword{datasets}
