% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications.R
\name{post_query}
\alias{post_query}
\title{Send query to Monday API}
\usage{
post_query(query, token = get_monday_token())
}
\arguments{
\item{query}{json (formatted according to \url{https://developer.monday.com/api-reference/docs/introduction-to-graphql})}

\item{token}{api access token}
}
\value{
result from \link[httr]{POST}
}
\description{
Send query to Monday API
}
