% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{process_roadmaps}
\alias{process_roadmaps}
\title{Perform processing of roadmaps to synchronize with tracking spreadsheets}
\usage{
process_roadmaps()
}
\description{
Perform processing of roadmaps to synchronize with tracking spreadsheets
}
