% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_sheets.R
\name{read_db_units}
\alias{read_db_units}
\title{Read the database of links for all units}
\usage{
read_db_units(id = getOption("gdrv_auto_env.URL_db_units"))
}
\arguments{
\item{id}{The URL for the googlesheet with the central DB of all units}
}
\value{
A tibble
}
\description{
Read the database of links for all units
}
