% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_sheets.R
\name{read_tracker_statuses}
\alias{read_tracker_statuses}
\title{Read the tasks for a single unit}
\usage{
read_tracker_statuses(url, sheet = "unit 1")
}
\arguments{
\item{url}{The URL for the googlesheet containing tasks for a single unit}

\item{sheet}{The name of the sheet to specify which unit, defaults to \code{"unit 1"}}
}
\value{
A tibble
}
\description{
Read the tasks for a single unit
}
