% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications.R
\name{send_notification}
\alias{send_notification}
\title{Send a notification on the item in Monday}
\usage{
send_notification(msg, item_id, user_id)
}
\arguments{
\item{msg}{contents of the notification}

\item{item_id}{id of the item for the notification to refer to}

\item{user_id}{id of the user to notify}
}
\description{
Send a notification on the item in Monday
}
