% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{set_loop_num_var}
\alias{set_loop_num_var}
\title{Set the loop number for the current logging session}
\usage{
set_loop_num_var()
}
\value{
a number, the current loop number
}
\description{
Set the loop number for the current logging session
}
