% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{stage_todo}
\alias{stage_todo}
\title{Record a todo in the log}
\usage{
stage_todo(action = "", url = "", loop_num = NULL)
}
\arguments{
\item{action}{Description of the todo item}

\item{url}{the URL for where to do the TODO}

\item{loop_num}{Number indicating which processing loop the todo message is associated with}
}
\description{
All changes to files in the C4R googledrive that are performed by this
package should be accompanied with a log entry, added via this function.
}
