% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_statuses.R
\name{title<-}
\alias{title<-}
\title{Set unit title of \code{statuses} object}
\usage{
title(x) <- value
}
\arguments{
\item{x}{object of type \code{statuses}}

\item{value}{new title (character)}
}
\value{
object of type \code{statuses}
}
\description{
Set unit title of \code{statuses} object
}
