% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_statuses.R
\name{title}
\alias{title}
\title{Get unit title of \code{statuses} object}
\usage{
title(x)
}
\arguments{
\item{x}{object of type \code{statuses}}
}
\value{
character (length 1)
}
\description{
Get unit title of \code{statuses} object
}
