% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications.R
\name{update_item}
\alias{update_item}
\title{Add an update to an existing item}
\usage{
update_item(item_body, item_id)
}
\arguments{
\item{item_body}{contents of the update}

\item{item_id}{id of the item to update (result from \code{\link[=create_item]{create_item()}})}
}
\value{
character (update_id of the new update)
}
\description{
Add an update to an existing item
}
