% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_sheets.R
\name{update_tracker_data}
\alias{update_tracker_data}
\title{Update the tracker data}
\usage{
update_tracker_data(tracker_dat, tracker_url, tracker_sheet)
}
\arguments{
\item{tracker_dat}{new data to put into the tracker}

\item{tracker_url}{url for the tracker}

\item{tracker_sheet}{sheet for the tracker}
}
\description{
Update the tracker data
}
