% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-functions.R
\name{URL_db_units}
\alias{URL_db_units}
\title{Default URL for the googlesheet with the central DB of all units}
\usage{
URL_db_units()
}
\value{
string with the URL for the
}
\description{
Default URL for the googlesheet with the central DB of all units
}
