% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications.R
\name{notify}
\alias{notify}
\title{Send notification}
\usage{
notify(msg, notify_text = "New Notification", to = "Hao Ye")
}
\arguments{
\item{msg}{contents of the notification}

\item{notify_text}{message on the notification itself}

\item{to}{a vector of names or emails}
}
\description{
Send notification
}
