% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications.R
\name{notify}
\alias{notify}
\title{Send notification}
\usage{
notify(to = "CENTER", msg)
}
\arguments{
\item{to}{either "CENTER" or "NIH" or a specific METER}

\item{msg}{contents of the notification}
}
\description{
Send notification
}
