% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications.R
\name{notify}
\alias{notify}
\title{Send notification}
\usage{
notify(to = "Hao Ye", msg, notify_text = "New Notification")
}
\arguments{
\item{to}{a vector of names or emails}

\item{msg}{contents of the notification}

\item{notify_text}{message on the notification itself}
}
\description{
Send notification
}
