% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit-data.R
\name{read_db_units}
\alias{read_db_units}
\title{Read the database of links for all units}
\usage{
read_db_units(id = URL_db_units())
}
\arguments{
\item{id}{The URL for the googlesheet with the master list of links for all units}
}
\value{
A tibble
}
\description{
Read the database of links for all units
}
